cleanup 3.0
